+++ atlasdb-config/src/main/java/com/palantir/atlasdb/factory/TransactionManagers.java
@@ -15,6 +15,7 @@
  */
 package com.palantir.atlasdb.factory;
 
+import java.util.Optional;
 import java.util.Set;
 import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.TimeUnit;
@@ -200,8 +201,16 @@ public final class TransactionManagers {
             LockServerOptions lockServerOptions,
             boolean allowHiddenTableAccess,
             Class<?> callingClass) {
-        return create(config, runtimeConfigSupplier, schemas, env, lockServerOptions, allowHiddenTableAccess,
-                UserAgents.fromClass(callingClass));
+        TransactionManagerOptions options = TransactionManagerOptions.builder()
+                .allowHiddenTableAccess(allowHiddenTableAccess)
+                .callingClass(callingClass)
+                .config(config)
+                .env(env)
+                .lockServerOptions(lockServerOptions)
+                .runtimeConfigSupplier(runtimeConfigSupplier)
+                .schemas(schemas)
+                .build();
+        return create(options);
     }
 
     public static SerializableTransactionManager create(
@@ -212,11 +221,25 @@ public final class TransactionManagers {
             LockServerOptions lockServerOptions,
             boolean allowHiddenTableAccess,
             String userAgent) {
+        TransactionManagerOptions options = TransactionManagerOptions.builder()
+                .allowHiddenTableAccess(allowHiddenTableAccess)
+                .config(config)
+                .env(env)
+                .lockServerOptions(lockServerOptions)
+                .runtimeConfigSupplier(optionalRuntimeConfigSupplier)
+                .schemas(schemas)
+                .userAgent(userAgent)
+                .build();
+        return create(options);
+    }
+
+    public static SerializableTransactionManager create(TransactionManagerOptions options) {
+        final AtlasDbConfig config = options.config();
         checkInstallConfig(config);
 
         AtlasDbRuntimeConfig defaultRuntime = AtlasDbRuntimeConfig.defaultRuntimeConfig();
         java.util.function.Supplier<AtlasDbRuntimeConfig> runtimeConfigSupplier =
-                () -> optionalRuntimeConfigSupplier.get().orElse(defaultRuntime);
+                () -> options.runtimeConfigSupplier().get().orElse(defaultRuntime);
 
         ServiceDiscoveringAtlasSupplier atlasFactory =
                 new ServiceDiscoveringAtlasSupplier(config.keyValueService(), config.leader(), config.namespace());
@@ -228,11 +251,11 @@ public final class TransactionManagers {
         LockAndTimestampServices lockAndTimestampServices = createLockAndTimestampServices(
                 config,
                 () -> runtimeConfigSupplier.get().timestampClient(),
-                env,
-                () -> LockServiceImpl.create(lockServerOptions),
+                options.env(),
+                () -> LockServiceImpl.create(options.lockServerOptions()),
                 atlasFactory::getTimestampService,
                 atlasFactory.getTimestampStoreInvalidator(),
-                userAgent);
+                options.derivedUserAgent());
 
         KeyValueService kvs = NamespacedKeyValueServices.wrapWithStaticNamespaceMappingKvs(rawKvs);
         kvs = ProfilingKeyValueService.create(kvs, config.getKvsSlowLogThresholdMillis());
@@ -240,12 +263,12 @@ public final class TransactionManagers {
                 new TimelockTimestampServiceAdapter(lockAndTimestampServices.timelock()));
         kvs = TracingKeyValueService.create(kvs);
         kvs = AtlasDbMetrics.instrument(KeyValueService.class, kvs,
-                MetricRegistry.name(KeyValueService.class, userAgent));
+                MetricRegistry.name(KeyValueService.class, options.derivedUserAgent()));
         kvs = ValidatingQueryRewritingKeyValueService.create(kvs);
 
         TransactionTables.createTables(kvs);
 
-        PersistentLockService persistentLockService = createAndRegisterPersistentLockService(kvs, env);
+        PersistentLockService persistentLockService = createAndRegisterPersistentLockService(kvs, options.env());
 
         TransactionService transactionService = TransactionServices.createTransactionService(kvs);
         ConflictDetectionManager conflictManager = ConflictDetectionManagers.create(kvs);
@@ -253,7 +276,7 @@ public final class TransactionManagers {
 
         Set<Schema> allSchemas = ImmutableSet.<Schema>builder()
                 .add(SweepSchema.INSTANCE.getLatestSchema())
-                .addAll(schemas)
+                .addAll(options.schemas())
                 .build();
         for (Schema schema : allSchemas) {
             Schemas.createTablesAndIndexes(schema, kvs);
@@ -263,7 +286,7 @@ public final class TransactionManagers {
         // TODO (jkong): Needs to be changed if/when we support dynamic table creation.
         LoggingArgs.hydrate(kvs.getMetadataForTables());
 
-        CleanupFollower follower = CleanupFollower.create(schemas);
+        CleanupFollower follower = CleanupFollower.create(options.schemas());
 
         Cleaner cleaner = new DefaultCleanerBuilder(
                 kvs,
@@ -286,14 +309,14 @@ public final class TransactionManagers {
                 conflictManager,
                 sweepStrategyManager,
                 cleaner,
-                allowHiddenTableAccess,
+                options.allowHiddenTableAccess(),
                 () -> runtimeConfigSupplier.get().transaction().getLockAcquireTimeoutMillis());
 
         PersistentLockManager persistentLockManager = new PersistentLockManager(
                 persistentLockService,
                 config.getSweepPersistentLockWaitMillis());
         initializeSweepEndpointAndBackgroundProcess(runtimeConfigSupplier,
-                env,
+                options.env(),
                 kvs,
                 transactionService,
                 sweepStrategyManager,